debug

+++ restore-archive.sh
@@ -1,6 +1,13 @@
+#!/bin/sh
+
+if [ $# -lt 1 ] ; then
+    echo 'Archive name required'
+    exit 1
+fi
+
 kubectl delete configmap archive -n turbobkup --ignore-not-found
 kubectl create configmap archive --from-literal=archive=$1 -n turbobkup
-kubectl apply -f testrestore.yaml -n turbobkup
+kubectl apply -f testrestorejob.yaml -n turbobkup
 # make sure the pod starts and runs
 count=0
 while [ "$(kubectl get pod -n turbobkup --no-headers | grep turbo-restore | awk '/Completed|Running/ {print $0}' )" = '' -a ${count} -lt 10 ] ; do
@@ -10,7 +17,7 @@ while [ "$(kubectl get pod -n turbobkup --no-headers | grep turbo-restore | awk 
 done
 if [ "$(kubectl get pod -n turbobkup --no-headers | grep turbo-restore | awk '/Completed|Running/ {print $0}' )" = '' ] ; then
     echo 'ERROR: Job Pod failed to start correctly'
-    kubectl delete -f testrestore.yaml
+    kubectl delete -f testrestorejob.yaml
 else
     echo '=============== wait for pod to finish ==============='
     count=0
@@ -23,7 +30,7 @@ else
     echo '=============== Logs ==============='
     if [ "$(kubectl get pod -n turbobkup --no-headers | grep Completed | grep turbo-restore | cut -f1 -d' ')" = '' ] ; then
         echo 'ERROR: Pod failed to complete successfully'
-        kubectl delete -f testrestore.yaml
+        kubectl delete -f testrestorejob.yaml
     fi
 fi
 kubectl delete configmap archive -n turbobkup
